Started work on class and methods to create cards using information
Need to sit down and think before acting though, so I'm gonna stop for today to do that.
Current plan is to spawn a card, edit it's information and then hand it off to the turn manager (not done). It needs to return the GameObject rather than a number really, so the turn manager can hand it to a player's deck.